+++ src/fuse_adapter.hpp
@@ -2,7 +2,9 @@
 #pragma once
 #include "filesystem.hpp"
 #include <filesystem>
+#include <functional>
 
 namespace macha {
-int run_fuse(FileSystem& fs, const std::filesystem::path& mount_path, bool allow_other = false);
+int run_fuse(FileSystem& fs, const std::filesystem::path& mount_path, bool allow_other = false,
+             std::function<void()> request_shutdown = {});
 }